+++ Stack/Opc.Ua.Core/Stack/Tcp/UaSCBinaryChannel.cs
@@ -257,7 +257,7 @@ namespace Opc.Ua.Bindings
                 }
             }
 
-            Utils.Trace((int)Utils.TraceMasks.Error, "{0}: Channel {1} - Duplicate sequence number: {2} <= {3}", context, this.ChannelId, sequenceNumber, m_remoteSequenceNumber);
+            Utils.Trace(Utils.TraceMasks.Error, "{0}: Channel {1} - Duplicate sequence number: {2} <= {3}", context, this.ChannelId, sequenceNumber, m_remoteSequenceNumber);
             return false;
         }
